requirements.txt
google-genai
deepgram-sdk
pydantic
python-dotenv
google
simpleaudio